requirement.txt
Flask
openai
mysql-connector-python
python-dotenv